xz: Use MSYS2/MinGW-w64 as the Windows build system
Adds common-scripts/msys2-env.bat to set the environment variables
necessary for MSYS2 to function correctly (in this scope, so far).

Also took the opportunity to be explcit with the flags to pass to
configure. We only want the libraries, and can't build the tools
since they depend on gettext which in turn depends on libiconv.