MSYS2: Add conversion script
Part of the real MSYS2 integration plan. This gets called from
build.sh and bld.bat to do all of the work of converting MSYS2
binary packages to conda packages.

I say all of the work, but actually the meta.yaml contains all
of the information needed already and it's a fairly simple job
at this stage.