MSYS2: Conversion script patching support
Uses a modified version of Anatoly Techtonik's patch.py

I added support for file deletion since that is an aspect
of patching that this stuff will use frequently.

For example, MSYS2's filesystem package needs a post-install
file deleted from it to prevent it from thinking it has
just freshly been installed.

Also another patch is applied to filesystem so that Windows
user temp is used as /tmp, a feature that came from the
Git-4-Windows project.