[WIP] r-base: Mid-debugging commit .. need to switch computers
Getting a crash 'Illegal Instruction' with boxplot(). I thought
it was the old, weird tcltk's fault, so I switched back to MSYS2's
but the crash is still there. Weird.

Well, on the +ve side, I did fix a problem with sqrt(NA_real_) which
needs to be sent upstream if it's not been fixed there already.